Add category search
Along with the ordinary search that finds terms among the
transliteration and definition, Dictionary search also includes
the ability to find entries of a specific language or part of
speech. This latter depends on the intersection being empty, i.e.:
not having a language and a part of speech with the same name.